requirements.txt
psycopg2-binary
python-dotenv
ijson
langchain
langchain-classic
langchain-ollama
langchain-chroma
chromadb
